Rename "history" to "context".
"Context" is a more general term used elsewhere for similar purposes.

Although contexts do not obviously have parents, neither did histories,
and no one should be surprised that these context objects happen to have a
chain of parents which are other context objects.

In the future, when Task becomes capable of things like cancellation and
mutual exclusion, the generality of "context" compared to "history" will
be even more valuable.