Avoid using WeakMap for Slot lookups.
The performance of WeakMap lookups is varied and unpredictable on
different systems and devices. An unguessable slot.id serves the same
purpose, without exposing any information or spoofing ability to parties
without a reference to the Slot. Of course, if you have a reference to a
Slot, it's up to you to avoid leaking its slot.id.