Consider undefined properties equivalent to missing properties. (#21)
Previously, an object `a` with a key `k` whose value is undefined (that
is, `a[k] === void 0`) would not be considered deeply equal to another
object `b` that is missing that key, but is otherwise deeply equal to the
first object, even though `a[k] === b[k]`.

This commit makes these two different kinds of undefined-ness equivalent,
by efficiently ignoring undefined-valued keys when comparing objects. This
allows more objects than before to be considered equivalent, but the
equal(a, b) function still obeys the rules of equivalence relations:
reflexivity, symmetry, and transitivity.

This change is similar in spirit to the way we relaxed the strict ===
requirement for function objects in #12: more functions can be considered
equivalent now, but the newly equivalent functions still obey the rules of
equality, and thus the change is logically valid, a decision we are free
to make but not forced to make.

The additional leniency will be useful in situations where (for example)
multiple options objects are combined by application code and then passed
to an API, perhaps using ...spread syntax or Object.assign, and the API
needs to decide if the new options are the same as previous options. This
change allows the application code to avoid worrying about the difference
between undefined and missing object properties, a discipline that can
unnecessarily complicate the way the options need to be combined.