Remove Node.js v10 from test rotation.
The `@wry/*` packages are all believed to work in Node 10, but that
Node.js version has been end-of-life'd (even for security updates) since
April 2021: https://endoflife.date/nodejs

Removing v10 should fix the tests in PR #230, which are broken because
Node 10 does not understand class property syntax in the `tests.cjs.js`
bundle.

The main `@wry/equality` library continues to be compiled to es2015,
eliminating class syntax, but the tests need to be compiled to esnext to
test generator and async function equality.

There may be a way to satisfy both of these constraints, but the easiest
solution right now is to avoid testing in Node 10.